Refactor transformers.js usage
