+++ default/config.conf
@@ -21,6 +21,8 @@ const extras = {
     classificationModel: 'Cohee/distilbert-base-uncased-go-emotions-onnx',
     // Image captioning model. HuggingFace ID of a model in ONNX format.
     captioningModel: 'Xenova/vit-gpt2-image-captioning',
+    // Feature extraction model. HuggingFace ID of a model in ONNX format.
+    embeddingModel: 'Xenova/all-mpnet-base-v2,
 };
 
 // Request overrides for additional headers
